[Fix] #117- 컬렉션 뷰 제거하고 다시 구현
